fix(hashline): isolate environment parser
